expired/lingua_1-main/lingua_1-main/electron_node/services/faster_whisper_vad/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
faster-whisper>=1.0.0
# GPU 支持：已配置为使用 GPU 加速
onnxruntime-gpu>=1.16.0
numpy>=1.24.0
soundfile>=0.12.0
scipy>=1.11.0
pydantic>=2.0.0
pydub>=0.25.1
pyogg>=0.6.12a1
# 注意：pydub 需要系统安装 ffmpeg（不是 Python 包）
# Windows: 下载 ffmpeg 并添加到 PATH
# Linux: sudo apt-get install ffmpeg
# macOS: brew install ffmpeg
# pyogg 用于直接解码原始 Opus 帧（优先使用）

